added movement to boss, tweaked impulses

+++ PerfectJazz/movement/cmp_move_sine.cpp
@@ -3,8 +3,8 @@
 void SineMovementComponent::update(double dt)
 {
 	_accumulation += dt;
-	float rotation = cos(_accumulation * _multiplier) * _maxAngle;
-	float moveRotation = cos(_accumulation);
+	float rotation = cosf(_accumulation * _multiplier) * _maxAngle;
+	float moveRotation = cosf(_accumulation);
 	_velocity = Vector2f(moveRotation, _velocity.y);
 	_parent->setRotation(rotation);	
 	MovementComponent::update(dt);
